Split annotation drawing methods into new module
